// src/app/precisa/maestros/ConfiguracionContratos/model/TipoOperacion.ts
export class tipoOperacion{

    constructor(){
    }

  TipoOperacionID:      number; 
  TIPOADMISIONID:       number;
  Persona:              number; 
  TPContratoID:         number; 
  IdListaBase:          number; 
  UneuNegocioId:        number; 
  TPAplicaFactor:       number; 
  FechaCreacion?:       Date; 
  FechaModificacion?:   Date; 
  UsuarioCreacion:      string; 
  UsuarioModificacion:  string; 
  IpCreacion:           string; 
  IpModificacion:       string; 
  TipEstado:            number; 
  TPFactor:             number; 
  TPMoneda:             string; 
  AplicaFormula:        number; 
  TipoPacienteId:       number; 
  IdSede:               number; 
  Observacion:          string; 
  FlatAprobacion:       number; 
  FlaCon:               number; 
  FlagRedondeo:         number; 
  FlagArchivo:          number; 
  IdSedeCliente:        number; 
  FlagAdelanto:         number; 
  FlagCortesia:         number; 
  MontoMinimo:          number; 
  AdmDescripcion:       string; 
  RucEmpresa:           string; 
  empresa:              string; 
  Nombre:               string; 
  EstadoDesc:           string; 
  SedDescripcion:       string; 
  FechaInicio?:         Date; 
  FechaTermino?:        Date; 
  Moneda:               string; 
  CodigoContrato:       string; 
  Monto:                number; 
  Descripcion:          string; 
  DescripcionLocal:     string; 
  Codigo:               string; 
  Convenio:             string; 
  DesSedeCliente:       string;
}